Split printer tests into per-topic modules
